(wip) adds tiny-cuda build process, although built for capacity 75 (mine is 61), it was able to try to find a kernel driver for running the code.
todo: move to a lmabdalab instance with a 75 capablity gpu (Quadro) and
build there, this addes the cli installation so if it builds we know
that we can run the scripts. I wanted to work on pushign output to a
remote drive but I tihnk it makes more sense to pull down the .plt and
dev the UI from there, for now. So this would semi-manual at the moment